requirements.txt
Flask>=2.0
qrcode[pil]
Pillow>=10.0
requests>=2.25
